; Conflicts:
;	db/schema.rb

+++ app/models/plan_land.rb
@@ -1,31 +1,35 @@
 # == Schema Information
 #
-# Table name: plan_lands # 作付計画
+# Table name: plan_lands
 #
 #  created_at             :datetime         not null
 #  updated_at             :datetime         not null
 #  land_id(土地)          :integer          not null
+#  user_id(利用者)        :integer          default(0), not null
 #  work_type_id(作業分類) :integer          not null
 #
 # Indexes
 #
-#  plan_lands_2nd  (land_id) UNIQUE
+#  plan_lands_2nd  (user_id,land_id) UNIQUE
 #
+
 class PlanLand < ApplicationRecord
   belongs_to :land
   belongs_to :work_type
+  belongs_to :user
 
-  def self.create_all(params)
-    PlanLand.delete_all
+  def self.create_all(user_id, params)
+    PlanLand.where(user_id: user_id).delete_all
     params.each do |param|
-      PlanLand.create(land_id: param[0], work_type_id: param[1]) if param[1].present?
+      PlanLand.create(user_id: user_id, land_id: param[0], work_type_id: param[1]) if param[1].present?
     end
   end
 
-  def self.clear_all(target)
-    PlanLand.delete_all
+  def self.clear_all(user_id, target)
+    PlanLand.where(user_id: user_id).delete_all
     Land.regionable.each do |land|
-      PlanLand.create(land_id: land.id, work_type_id: land.cost(target).work_type_id)
+      land_cost = land.cost(target)
+      PlanLand.create(user_id: user_id, land_id: land.id, work_type_id: land_cost.work_type_id) if land_cost
     end
   end
 end
